==> how-to-win-a-data-science-competition/13_final-project/FinalProject_FutureSalePredict/code/01_first_view.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import pandas as  pd
import numpy as np
import warnings
import itertools
import xgboost as xgb
warnings.filterwarnings('ignore')
kernel_with_output = True

df_items=pd.read_csv("../data/items.csv")

df_items.info()
df_items.head()

df_categories = pd.read_csv("../data/item_categories.csv")
df_categories.info()
df_categories.head()

df_shops = pd.read_csv("../data/shops.csv")
df_shops.info()
df_shops.head()

df_sales = pd.read_csv("../data/sales_train.csv.gz")
df_sales.info()
df_sales.head()

df_test = pd.read_csv("../data/test.csv.gz")
df_test.info()
df_test.head()

temp = pd.read_csv("../data/sample_submission.csv.gz")
temp.head()

### check null

df_items.isnull().sum()
df_categories.isnull().sum()
df_shops.isnull().sum()
df_sales.isnull().sum()
df_test.isnull().sum();

there is no NA in any tables here

df_sales.dtypes

df_sales["date"]=pd.to_datetime(df_sales["date"],format="%d.%m.%Y")
df_sales["month"]=df_sales["date"].dt.month
df_sales["year"] = df_sales["date"].dt.year

df_sales.info()
df_sales.loc[(df_sales.year==2015) & (df_sales.month==10), "date_block_num"].unique()

#### aggregation

##### originally, item_cnt_day almost 1 (untill 75 percentile = 1), max = 2169, there are outliers but we do not want to delete them, so we clip them to 20. Then we make any possible combination of shop_id, item_id, data_block_num. After that we fill in the places where these combination has values, otherwise fill 0.

df_sales["item_cnt_day"] = df_sales["item_cnt_day"].clip(0,20)

group_by_columns = ["shop_id", "item_id", "date_block_num"]
grid = [] 
for block_num in df_sales['date_block_num'].unique():
    block = df_sales[df_sales['date_block_num']==block_num]
    cur_shops = block['shop_id'].unique()
    cur_items = block['item_id'].unique()
    grid.append(np.array(list(itertools.product(*[cur_shops, cur_items, [block_num]])), dtype='int32'))
grid = pd.DataFrame(np.vstack(grid), columns = group_by_columns)
df_temp=df_sales.groupby(group_by_columns,as_index=False).agg({"item_cnt_day":"sum", "item_price":"mean"})

df_temp.info()
df_temp.head()

all_data = pd.merge(grid, df_temp, how="left", on = group_by_columns).fillna(0)
all_data.sort_values(["date_block_num", "shop_id", "item_id"], inplace = True)
all_data.rename({"item_cnt_day":"item_cnt_month"}, inplace = True, axis = "columns")
all_data = pd.merge(all_data, df_items[["item_id", "item_category_id"]], on = "item_id")
all_data["item_cnt_month"] = all_data["item_cnt_month"].clip(0, 20)
display(all_data.head(10))
all_data.shape

all_data.to_csv("../results/01/all_data.csv")

### Previous Value Benchmark

if kernel_with_output:
    train_subset = all_data[all_data.date_block_num==33]
    train_subset = train_subset.groupby(["shop_id", "item_id"]).agg({"item_cnt_month":"sum"}).reset_index()
    train_subset.head(5)

if kernel_with_output:
    merged = df_test.merge(train_subset, on =["shop_id", "item_id"], how = "left")[["ID", "item_cnt_month"]]

merged["item_cnt_month"] = merged["item_cnt_month"].fillna(0).clip(0,20)

merged.set_index("ID").to_csv("../results/01/2.csv")

#### this score 1.167777
