==> how-to-win-a-data-science-competition/13_final-project/FinalProject_FutureSalePredict/code/03_Xgboost_baseline.ipynb <==
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

import pandas as  pd
import numpy as np
import warnings
import itertools
import xgboost as xgb
from numpy import loadtxt
warnings.filterwarnings('ignore')
kernel_with_output = True

if kernel_with_output:
    df_items=pd.read_csv("../data/items.csv")
    df_categories = pd.read_csv("../data/item_categories.csv")
    df_shops = pd.read_csv("../data/shops.csv")
    df_sales = pd.read_csv("../data/sales_train.csv.gz")
    df_test = pd.read_csv("../data/test.csv.gz")
    temp = pd.read_csv("../data/sample_submission.csv.gz")

df_sales["date"]=pd.to_datetime(df_sales["date"],format="%d.%m.%Y")
df_sales["month"]=df_sales["date"].dt.month
df_sales["year"] = df_sales["date"].dt.year

all_data = pd.read_csv("../results/all_data.csv")

### XGBoost baseline

if kernel_with_output:
    # Extract features and target we want
    baseline_features = ['shop_id', 'item_id', 'item_category_id', 'date_block_num', 'item_cnt_month']
    train = all_data[baseline_features]
    # Remove pandas index column
    train = train.set_index('shop_id')
    train["item_cnt_month"] = train["item_cnt_month"].astype(int)
    train['item_cnt_month'] = train["item_cnt_month"].fillna(0).clip(0,20)
    # Save train set to file
    train.to_csv('../results/train_baseline.csv')


if kernel_with_output:
    dataset = loadtxt('../results/train.csv', delimiter="," ,skiprows=1, dtype = int)
    # trainx's columns: 'shop_id', 'item_id', 'item_category_id', 'date_block_num'
    trainx = dataset[:, 0:4] # without header 
    trainy = dataset[:, 4]

    # Make test_dataset pandas data frame, add category id from df_items, and fix date block num 33
    merged_test = pd.merge(df_test, df_items, on = ['item_id'])[['shop_id','item_id','item_category_id']]
    merged_test['date_block_num'] = 33
    merged_test.set_index('shop_id')
    

if kernel_with_output:
    model = xgb.XGBRegressor(max_depth = 10, min_child_weight=0.5, subsample = 1, eta = 0.3, num_round = 50, seed = 1)
    model.fit(trainx, trainy, eval_metric='rmse')

 preds = model.predict(merged_test.values)

if kernel_with_output:
    df = pd.DataFrame(preds, columns = ['item_cnt_month'])
    df['ID'] = df.index
    df = df.set_index('ID')
    df.head(10)
    df.to_csv('../results/03.csv')

#### result 1.35
