==> how-to-win-a-data-science-competition/13_final-project/FinalProject_FutureSalePredict/code/06_FeaturesEngineerAndEnsemble.ipynb <==
import pandas as  pd
import numpy as np
import warnings
import itertools
import xgboost as xgb
from tqdm import tqdm
from numpy import loadtxt
import time
import gc
warnings.filterwarnings('ignore')
kernel_with_output = True
np.random.seed(10)

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"
pd.set_option('display.max_rows', 231)
pd.set_option('display.max_columns', 100)

Validation = False
start_time = time.time()

def downcast_dtypes(df):
    float_cols = [c for c in df if df[c].dtype == "float64"]
    int_cols =   [c for c in df if df[c].dtype in ["int64", "int32"]]
    df[float_cols] = df[float_cols].astype(np.float32)
    df[int_cols]   = df[int_cols].astype(np.int16)
    return df

if kernel_with_output:
    item =pd.read_csv("../data/items.csv")
    item_cat = pd.read_csv("../data/item_categories.csv")
#     df_shops = pd.read_csv("../data/shops.csv")
    sale_train = pd.read_csv("../data/sales_train.csv.gz")
    test = pd.read_csv("../data/test.csv.gz")
    temp = pd.read_csv("../data/sample_submission.csv.gz")


sale_train.loc[2909818] # id = 2909818 seems outliers

sale_train['item_price'][2909818] = sale_train[(sale_train['shop_id'] ==12) & (sale_train['item_id'] == 11373) & (sale_train['date_block_num'] == 33)]['item_price'].median()
sale_train['item_cnt_day'][2909818] = round(sale_train[(sale_train['shop_id'] ==12) & (sale_train['item_id'] == 11373) & (sale_train['date_block_num'] == 33)]['item_cnt_day'].median())

sale_train.loc[2909818] 

sale_train.loc[885138] # the price seems too high

sale_train['item_price'][885138] = sale_train[(sale_train['item_id'] == 11365) & (sale_train['shop_id'] ==12) & (sale_train['date_block_num'] == 8)]['item_price'].median()

sale_train.loc[885138] # the price seems too high

sale_train.shape

test_nrow = test.shape[0]

sale_train = sale_train.merge(test[['shop_id']].drop_duplicates(), how = 'inner') 
sale_train['date'] = pd.to_datetime(sale_train['date'], format = '%d.%m.%Y')

After merge, we delete 522603 records from sale_train. The reason is we want drop all transactions belong to shops in the train set but not in the test. 

sale_train.info()

There are 4 steps:

(1) Aggregate data

(2) Add item/shop pair mean-encoding

(3) First-level model

(4) Ensembling

### 1. Aggregate data 

grid = []
for block_num in sale_train['date_block_num'].unique():
    cur_shops = sale_train[sale_train['date_block_num']==block_num]['shop_id'].unique()
    cur_items = sale_train[sale_train['date_block_num']==block_num]['item_id'].unique()
    grid.append(np.array(list(itertools.product(*[cur_shops, cur_items, [block_num]])),dtype='int32'))

#turn the grid into pandas dataframe
index_cols = ['shop_id', 'item_id', 'date_block_num']
grid = pd.DataFrame(np.vstack(grid), columns = index_cols,dtype=np.int32)
print('%0.2f min: Finish creating the grid'%((time.time() - start_time)/60))

index_cols = ['shop_id', 'item_id', 'date_block_num']
sale_train['item_cnt_day'] = sale_train['item_cnt_day'].clip(0,20)
gb_cnt = sale_train.groupby(index_cols)['item_cnt_day'].agg(['sum']).reset_index().rename(columns = {'sum': 'item_cnt_month'})
gb_cnt['item_cnt_month'] = gb_cnt['item_cnt_month'].clip(0,20).astype(np.int)
#join aggregated data to the grid
train = pd.merge(grid,gb_cnt,how='left',on=index_cols).fillna(0)
train['item_cnt_month'] = train['item_cnt_month'].astype(int)

train.info()

train = train.merge(item[['item_id', 'item_category_id']], on = ['item_id'], how = 'left')
test = test.merge(item[["item_id", "item_category_id"]], on = ["item_id"], how="left")

item_cat.head()
item_cat.info()

item_cat = pd.read_csv('../data/item_categories.csv')

l_cat = list(item_cat.item_category_name)

for ind in range(0,1):

    l_cat[ind] = 'PC Headsets / Headphones'

for ind in range(1,8):

    l_cat[ind] = 'Access'

l_cat[8] = 'Tickets (figure)'

l_cat[9] = 'Delivery of goods'

for ind in range(10,18):

    l_cat[ind] = 'Consoles'

for ind in range(18,25):

    l_cat[ind] = 'Consoles Games'

l_cat[25] = 'Accessories for games'

for ind in range(26,28):

    l_cat[ind] = 'phone games'

for ind in range(28,32):

    l_cat[ind] = 'CD games'

for ind in range(32,37):

    l_cat[ind] = 'Card'

for ind in range(37,43):

    l_cat[ind] = 'Movie'

for ind in range(43,55):

    l_cat[ind] = 'Books'

for ind in range(55,61):

    l_cat[ind] = 'Music'

for ind in range(61,73):

    l_cat[ind] = 'Gifts'

for ind in range(73,79):

    l_cat[ind] = 'Soft'

for ind in range(79,81):

    l_cat[ind] = 'Office'

for ind in range(81,83):

    l_cat[ind] = 'Clean'

l_cat[83] = 'Elements of a food'


from sklearn import preprocessing
lb = preprocessing.LabelEncoder()
item_cat['item_cat_id_fix'] = lb.fit_transform(l_cat)
train = train.merge(item_cat[['item_cat_id_fix', 'item_category_id']], on = ['item_category_id'], how = 'left')
test = test.merge(item_cat[['item_cat_id_fix', 'item_category_id']], on = ['item_category_id'], how = 'left')

item_cat.head(50)

train = downcast_dtypes(train)
test = downcast_dtypes(test)

train.info()
test.info()

import pickle as pkl

pkl.dump(train, open("../results/06/train.pkl","wb"))
pkl.dump(test, open("../results/06/test.pkl","wb"))

### 2. Add item/shop pair mean-encodings

##### 2.1 Combine trainset and testset

print('%0.2f min: Start combining data'%((time.time() - start_time)/60))
if Validation == False:
    test['date_block_num'] = 34
    all_data = pd.concat([train, test], axis = 0)
    all_data = all_data.drop(columns = ['ID'])
else:
    all_data = train

all_data.shape
all_data.isnull().sum() # test has only 5 cols, so when merge, these cols will be na

all_data = downcast_dtypes(all_data)

all_data.head()

##### 2.2 Creating item/shop pair lags lag-based featuers

print('%0.2f min: Start adding lag-based feature'%((time.time() - start_time)/60))
index_cols = ['shop_id', 'item_id', 'item_category_id', 'item_cat_id_fix', 'date_block_num']
cols_to_rename = list(all_data.columns.difference(index_cols))
print(cols_to_rename)
shift_range = [1, 2, 3, 4, 12]

for month_shift in tqdm(shift_range):
    train_shift = all_data[index_cols + cols_to_rename].copy()
    train_shift['date_block_num'] = train_shift['date_block_num'] + month_shift
    foo = lambda x: '{}_lag_{}'.format(x, month_shift) if x in cols_to_rename else x
    train_shift = train_shift.rename(columns=foo)
    all_data = pd.merge(all_data, train_shift, on=index_cols, how='left').fillna(0)
del train_shift
gc.collect()

all_data = all_data[all_data['date_block_num'] >= 12] # Don't use old data from year 2013
lag_cols = [col for col in all_data.columns if col[-1] in [str(item) for item in shift_range]]
all_data = downcast_dtypes(all_data)
print('%0.2f min: Finish generating lag features'%((time.time() - start_time)/60))

lag_cols

all_data.shape

##### 2.3 Creating date features

dates_train = sale_train[['date', 'date_block_num']].drop_duplicates()

# choose 22 because date_block_num is Nov 2014, We will make this data to be Nov 2015
dates_test = dates_train[dates_train['date_block_num'] == 34-12] 

dates_train.head()
dates_train.info()
dates_train.date_block_num.value_counts()
dates_test.head()
dates_test.info()
dates_test.date_block_num.value_counts()

dates_test['date_block_num'] = 34
dates_test['date'] = dates_test['date'] + pd.DateOffset(years=1) # increase 1 year

dates_test.head()
dates_test.info()
dates_test.date_block_num.value_counts()

dates_all = pd.concat([dates_train, dates_test])

dates_all.date_block_num.value_counts()

dates_all['dow'] = dates_all['date'].dt.dayofweek
dates_all['year'] = dates_all['date'].dt.year
dates_all['month'] = dates_all['date'].dt.month

dates_all = pd.get_dummies(dates_all, columns=['dow'])
dow_col = ['dow_' + str(x) for x in range(7)]

dow_col

date_features = dates_all.groupby(['year', 'month', 'date_block_num'])[dow_col].agg('sum').reset_index()

dates_all.groupby(['year', 'month', 'date_block_num'])[dow_col].agg('sum').reset_index()

date_features['days_of_month'] = date_features[dow_col].sum(axis=1)

date_features["days_of_month"].describe()

date_features['year'] = date_features['year'] - 2013 # We already choose data from 2013, so we do it
# to make year start from 0. Maybe this helps to make the value of "year" smaller

date_features = date_features[['month', 'year', 'days_of_month', 'date_block_num']]

all_data = all_data.merge(date_features, on = 'date_block_num', how = 'left')

all_data.head()

date_columns = date_features.columns.difference(set(index_cols))

date_columns

print('%0.2f min: Start getting date features'%((time.time() - start_time)/60))
dates_train = sale_train[['date', 'date_block_num']].drop_duplicates()
dates_test = dates_train[dates_train['date_block_num'] == 34-12]
dates_test['date_block_num'] = 34
dates_test['date'] = dates_test['date'] + pd.DateOffset(years=1) # increase 1 year

dates_all = pd.concat([dates_train, dates_test])
dates_all['dow'] = dates_all['date'].dt.dayofweek
dates_all['year'] = dates_all['date'].dt.year
dates_all['month'] = dates_all['date'].dt.month
dates_all = pd.get_dummies(dates_all, columns=['dow'])
dow_col = ['dow_' + str(x) for x in range(7)]
date_features = dates_all.groupby(['year', 'month', 'date_block_num'])[dow_col].agg('sum').reset_index()
date_features['days_of_month'] = date_features[dow_col].sum(axis=1)
date_features['year'] = date_features['year'] - 2013
date_features = date_features[['month', 'year', 'days_of_month', 'date_block_num']]
all_data = all_data.merge(date_features, on = 'date_block_num', how = 'left')
date_columns = date_features.columns.difference(set(index_cols))
print('%0.2f min: Finish getting date features'%((time.time() - start_time)/60))

#### 2.4 Scale feature columns

from sklearn.preprocessing import StandardScaler

train = all_data[all_data["date_block_num"]!= all_data["date_block_num"].max()]
test = all_data[all_data["date_block_num"]== all_data["date_block_num"].max()]

test.head()

sc = StandardScaler()

to_drop_cols = ["date_block_num"]

index_cols

feature_columns = list(set(lag_cols + index_cols + list(date_columns)).difference(to_drop_cols))

feature_columns

train.rename(columns = {"month_x":"month", "year_x":"year", "days_of_month_x":"days_of_month"}, inplace = True)
test.rename(columns = {"month_x":"month", "year_x":"year", "days_of_month_x":"days_of_month"}, inplace = True)

train.drop(["month_y", "year_y", "days_of_month_y"], axis = 1, inplace = True)
test.drop(["month_y", "year_y", "days_of_month_y"], axis = 1, inplace = True)

test.info()

train[feature_columns] = sc.fit_transform(train[feature_columns])
test[feature_columns] = sc.fit_transform(test[feature_columns])

all_data = pd.concat([train, test], axis = 0)
all_data = downcast_dtypes(all_data)

all_data.to_csv("../results/06/step_2.csv")

gc.collect()
print('%0.2f min: Finish scaling features'%((time.time() - start_time)/60))

### 3. First-level model

###### Save `date_block_num`, as we can't use them as features, but will need them to split the dataset into parts

dates = all_data["date_block_num"]
last_block = dates.max() # last_block 34

print("%0.2f min: Start training First level models" %((time.time() - start_time)/60))

start_first_level_total = time.perf_counter()
scoringMethod = 'r2'
num_first_level_models = 3 

from sklearn.metrics import mean_squared_error
from math import sqrt

# Train meta-features M = 15 (12 + 15 = 27)

months_to_generate_meta_features = range(27,last_block +1)
mask = dates.isin(months_to_generate_meta_features)
Target = 'item_cnt_month'
y_all_level2 = all_data[Target][mask].values # choose item_cnt_month where date_block_num from 27..34
X_all_level2 = np.zeros([y_all_level2.shape[0], num_first_level_models])

slice_start = 0

SEED = 0
for cur_block_num in tqdm(months_to_generate_meta_features):
    print('-' * 50)
    print('Start training for month%d'% cur_block_num)
    start_cur_month = time.perf_counter()
    cur_X_train = all_data.loc[dates <  cur_block_num][feature_columns]
    cur_X_test =  all_data.loc[dates == cur_block_num][feature_columns]
    cur_y_train = all_data.loc[dates <  cur_block_num, Target].values
    cur_y_test =  all_data.loc[dates == cur_block_num, Target].values
    # Create Numpy arrays of train, test and target dataframes to feed into models
    train_x = cur_X_train.values
    train_y = cur_y_train.ravel()
    test_x = cur_X_test.values
    test_y = cur_y_test.ravel()
    
    preds = []
    from sklearn.linear_model import (LinearRegression, SGDRegressor)
    import lightgbm as lgb
    sgdr= SGDRegressor(
        penalty = 'l2' ,
        random_state = SEED )
    lgb_params = {
        'feature_fraction': 0.75,
        'metric': 'rmse',
        'nthread':1,
        'min_data_in_leaf': 2**7,
        'bagging_fraction': 0.75,
        'learning_rate': 0.03,
        'objective': 'mse',
        'bagging_seed': 2**7,
        'num_leaves': 2**7,
        'bagging_freq':1,
        'verbose':0}
    estimators = [sgdr]
    for estimator in estimators:
        print('Training Model %d: %s'%(len(preds), estimator.__class__.__name__))
        start = time.perf_counter()
        estimator.fit(train_x, train_y)
        pred_test = estimator.predict(test_x)
        preds.append(pred_test)

        pred_train = estimator.predict(train_x)
        print('Train RMSE for %s is %f' % (estimator.__class__.__name__, sqrt(mean_squared_error(cur_y_train, pred_train))))
        print('Test RMSE for %s is %f' % (estimator.__class__.__name__, sqrt(mean_squared_error(cur_y_test, pred_test))))

        run = time.perf_counter() - start
        print('{} runs for {:.2f} seconds.'.format(estimator.__class__.__name__, run))
        print()
#         import pickle
#         pickle.dump(estimator, open(filename, 'wb'))

    print('Training Model %d: %s'%(len(preds), 'lightgbm'))
    start = time.perf_counter()
    estimator = lgb.train(lgb_params, lgb.Dataset(train_x, label=train_y), 300)
    pred_test = estimator.predict(test_x)
    preds.append(pred_test)

    pred_train = estimator.predict(train_x)
    print('Train RMSE for %s is %f' % ('lightgbm', sqrt(mean_squared_error(cur_y_train, pred_train))))
    print('Test RMSE for %s is %f' % ('lightgbm', sqrt(mean_squared_error(cur_y_test, pred_test))))

    run = time.perf_counter() - start
    print('{} runs for {:.2f} seconds.'.format('lightgbm', run))
    print()

    print('Training Model %d: %s'%(len(preds), 'keras'))
    start = time.perf_counter()
    from keras.models import Sequential
    from keras.layers import Dense
    from keras.wrappers.scikit_learn import KerasRegressor

    def baseline_model():
        # create model
        model = Sequential()
        model.add(Dense(20, input_dim=train_x.shape[1], kernel_initializer='uniform', activation='softplus'))
        model.add(Dense(1, kernel_initializer='uniform', activation = 'relu'))
        # Compile model
        model.compile(loss='mse', optimizer='Nadam', metrics=['mse'])
        # model.compile(loss='mean_squared_error', optimizer='adam')
        return model

    estimator = KerasRegressor(build_fn=baseline_model, verbose=1, epochs=5, batch_size = 55000)
    estimator.fit(train_x, train_y)
    pred_test = estimator.predict(test_x)
    preds.append(pred_test)
    run = time.perf_counter() - start
    print('{} runs for {:.2f} seconds.'.format('lightgbm', run))

    cur_month_run_total = time.perf_counter() - start_cur_month
    print('Total running time was {:.2f} minutes.'.format(cur_month_run_total/60))
    print('-' * 50)

    slice_end = slice_start + cur_X_test.shape[0]
    X_all_level2[ slice_start : slice_end , :] = np.c_[preds].transpose()
    slice_start = slice_end

# Split train and test
test_nrow = len(preds[0])
X_train_level2 = X_all_level2[ : -test_nrow, :]
X_test_level2 = X_all_level2[ -test_nrow: , :]
y_train_level2 = y_all_level2[ : -test_nrow]
y_test_level2 = y_all_level2[ -test_nrow : ]
print('%0.2f min: Finish training First level models'%((time.perf_counter() - start_first_level_total)/60))

### 4. Ensembling

pred_list = {}

# A. Second level learning model via linear regression
print('Training Second level learning model via linear regression')
from sklearn.linear_model import (LinearRegression, SGDRegressor)
lr = LinearRegression()
lr.fit(X_train_level2, y_train_level2)
# Compute R-squared on the train and test sets.
# print('Train R-squared for %s is %f' %('test_preds_lr_stacking', sqrt(mean_squared_error(y_train_level2, lr.predict(X_train_level2)))))
test_preds_lr_stacking = lr.predict(X_test_level2)
train_preds_lr_stacking = lr.predict(X_train_level2)
print('Train R-squared for %s is %f' %('train_preds_lr_stacking', sqrt(mean_squared_error(y_train_level2, train_preds_lr_stacking))))
pred_list['test_preds_lr_stacking'] = test_preds_lr_stacking
if Validation:
    print('Test R-squared for %s is %f' %('test_preds_lr_stacking', sqrt(mean_squared_error(y_test_level2, test_preds_lr_stacking))))

# B. Second level learning model via SGDRegressor
print('Training Second level learning model via SGDRegressor')
sgdr= SGDRegressor(
    penalty = 'l2' ,
    random_state = SEED )
sgdr.fit(X_train_level2, y_train_level2)
# Compute R-squared on the train and test sets.
# print('Train R-squared for %s is %f' %('test_preds_lr_stacking', sqrt(mean_squared_error(y_train_level2, lr.predict(X_train_level2)))))
test_preds_sgdr_stacking = sgdr.predict(X_test_level2)
train_preds_sgdr_stacking = sgdr.predict(X_train_level2)
print('Train R-squared for %s is %f' %('train_preds_lr_stacking', sqrt(mean_squared_error(y_train_level2, train_preds_sgdr_stacking))))
pred_list['test_preds_sgdr_stacking'] = test_preds_sgdr_stacking
if Validation:
    print('Test R-squared for %s is %f' %('test_preds_sgdr_stacking', sqrt(mean_squared_error(y_test_level2, test_preds_sgdr_stacking))))
print('%0.2f min: Finish training second level model'%((time.time() - start_time)/60))

# Submission -------------------------------------------------------------------
if not Validation:
    submission = pd.read_csv('../data/sample_submission.csv')
    ver = 6
    for pred_ver in ['lr_stacking', 'sgdr_stacking']:
        print(pred_list['test_preds_' + pred_ver].clip(0,20).mean())
        submission['item_cnt_month'] = pred_list['test_preds_' + pred_ver].clip(0,20)
        submission[['ID', 'item_cnt_month']].to_csv("../results/05.csv", index = False)
print('%0.2f min: Finish running scripts'%((time.time() - start_time)/60))

!shutdown -s
